fix(embeds): Allow color override in build_info_embed
The `build_info_embed` function was being called with a `color` argument from the refinery command, but the function did not accept this argument. This caused an error for users who had not yet produced any melange.

This change modifies the `build_info_embed` function to accept an optional `color` argument, which is then passed to the underlying `build_status_embed` function. This allows the color of the info embed to be customized while maintaining the default blue color for other uses.